Restructures hero component and layout
Refactors the hero component to use slots for content injection,
improving flexibility and customization.

Moves hero content, previously defined as props on the
article layout, into the hero component's default slot. This allows
for more complex hero designs and better content management.

Renames the 'footer' slot in the hero component to 'bottom' for
clarity and consistency.